torch/requirements.txt
numpy
scipy
scikit-learn
opencv-python
opencv-contrib-python
torch
torchvision
torchsummary
einops
matplotlib
tqdm
pillow
mnn
onnx
onnxruntime
